requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
prometheus-client==0.19.0
structlog==23.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
nltk==3.8.1
scikit-learn==1.3.2
transformers==4.36.0
torch==2.1.0
numpy==1.24.4
pandas==2.1.4
